requirements.txt
python-telegram-bot>=13.0
psycopg2-binary>=2.9.3
python-dateutil>=2.8.0